# ARROW-10468: [C++][Compute] Provide KernelExecutor instead of FunctionExecutor