--- a/cpp/src/arrow/compute/exec_internal.h
+++ b/cpp/src/arrow/compute/exec_internal.h
@@ -102,9 +102,9 @@ class DatumAccumulator : public ExecListener {
 /// inputs will be split into non-chunked ExecBatch values for execution
 Status CheckAllValues(const std::vector<Datum>& values);
 
-class ARROW_EXPORT FunctionExecutor {
+class ARROW_EXPORT KernelExecutor {
  public:
-  virtual ~FunctionExecutor() = default;
+  virtual ~KernelExecutor() = default;
 
   /// XXX: Better configurability for listener
   /// Not thread-safe
@@ -115,9 +115,12 @@ class ARROW_EXPORT FunctionExecutor {
   virtual Datum WrapResults(const std::vector<Datum>& args,
                             const std::vector<Datum>& outputs) = 0;
 
-  static Result<std::unique_ptr<FunctionExecutor>> Make(ExecContext* ctx,
-                                                        const Function* func,
-                                                        const FunctionOptions* options);
+  static Result<std::unique_ptr<KernelExecutor>> MakeScalar(ExecContext* ctx,
+                                                            KernelInitArgs);
+  static Result<std::unique_ptr<KernelExecutor>> MakeVector(ExecContext* ctx,
+                                                            KernelInitArgs);
+  static Result<std::unique_ptr<KernelExecutor>> MakeScalarAggregate(ExecContext* ctx,
+                                                                     KernelInitArgs);
 };
 
 /// \brief Populate validity bitmap with the intersection of the nullity of the