… cpp/src/arrow/compute/CMakeLists.txt
… cpp/src/arrow/compute/function.h
… cpp/src/arrow/compute/function_benchmark.cc
… cpp/src/arrow/compute/function_test.cc